Track just the latest fingerprints
Each feed check, update self.fingerprints to be a set of all
fingerprints in the feed.

I think the previous way I was doing it was a holdover from how I did it
in riverpy which just didn't make much sense here.

Next up is to trim self.timestamps so it's never longer than
self.window.