Don't skip updates when the feed has just been added
When the JSON file already exists upon startup, all updates from the
initial check are skipped and only new updates from then on are
written out.

Previously, this would also be the case for feeds added to the feed
list.

With this commit, updates from newly added feeds appear whether the
JSON file existed already or not.